+++ microservices/butakero_bot/internal/infrastructure/adapters/audio_api_client_test.go
@@ -1,4 +1,6 @@
-package client
+//go:build !integration
+
+package adapters
 
 import (
 	"context"
